Fix WD1770 formatting issues
Fix some formatting issiues for drives using the WD1770 caused by
over-eager casting of data values that removed the occasional flag bit
in bit 8.

git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@35098 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/drive/iec/wd1770.c
@@ -153,7 +153,7 @@ struct wd1770_s {
     fd_drive_t *fdd;
     int step;
     int byte_count;
-    int tmp;
+    unsigned int tmp;
     int direction;
     int clock_frequency; /* MHz of main CPU*/
 
@@ -208,7 +208,7 @@ void wd1770_shutdown(wd1770_t *drv)
 /* Execute microcode */
 static void wd1770_execute(wd1770_t *drv)
 {
-    int res;
+    unsigned int res;
 
     for (;; ) {
         switch (drv->type) {
@@ -361,7 +361,7 @@ static void wd1770_execute(wd1770_t *drv)
                             drv->step--;
                             continue;
                         }
-                        drv->crc = fdd_crc(drv->crc, (uint8_t)res);
+                        drv->crc = fdd_crc(drv->crc, res);
                         if (--drv->byte_count) {
                             continue;
                         }
@@ -466,7 +466,7 @@ static void wd1770_execute(wd1770_t *drv)
                         if (drv->byte_count == 3) {
                             drv->tmp = res;
                         }
-                        drv->crc = fdd_crc(drv->crc, (uint8_t)res);
+                        drv->crc = fdd_crc(drv->crc, res);
                         if (--drv->byte_count) {
                             continue;
                         }
@@ -500,12 +500,12 @@ static void wd1770_execute(wd1770_t *drv)
                                 if (!drv->sync) {
                                     drv->crc = 0xffff;
                                 }
-                                drv->crc = fdd_crc(drv->crc, (uint8_t)res);
+                                drv->crc = fdd_crc(drv->crc, res);
                                 drv->sync = (res == 0x1a1);
                                 continue;
                             }
                         }
-                        drv->crc = fdd_crc(drv->crc, (uint8_t)res);
+                        drv->crc = fdd_crc(drv->crc, res);
                         drv->status |= ((res & 0xff) == 0xf8) ? WD_RT : 0;
                         drv->byte_count = (128 << drv->tmp) + 2;
                         drv->step++;
@@ -520,7 +520,7 @@ static void wd1770_execute(wd1770_t *drv)
                             drv->status |= (drv->status & WD_DRQ) ? WD_LD : WD_DRQ;
                             drv->data = res;
                         }
-                        drv->crc = fdd_crc(drv->crc, (uint8_t)res);
+                        drv->crc = fdd_crc(drv->crc, res);
                         if (--drv->byte_count) {
                             continue;
                         }
@@ -562,8 +562,8 @@ static void wd1770_execute(wd1770_t *drv)
                             continue;
                         }
                         res = ((drv->cmd & WD_A) ? 0xf8 : 0xfb) | (drv->dden ? 0x100 : 0);
-                        fdd_write(drv->fdd, (uint8_t)res);
-                        drv->crc = fdd_crc(drv->crc, (uint8_t)res);
+                        fdd_write(drv->fdd, res);
+                        drv->crc = fdd_crc(drv->crc, res);
                         drv->byte_count = (128 << drv->tmp) + 3;
                         drv->step++;
                         /* fall through */
@@ -577,10 +577,10 @@ static void wd1770_execute(wd1770_t *drv)
                                 fdd_write(drv->fdd, 0xff);
                                 break;
                             case 1:
-                                fdd_write(drv->fdd, (uint8_t)(drv->crc & 0xff));
+                                fdd_write(drv->fdd, drv->crc & 0xff);
                                 continue;
                             case 2:
-                                fdd_write(drv->fdd, (uint8_t)(drv->crc >> 8));
+                                fdd_write(drv->fdd, drv->crc >> 8);
                                 continue;
                             default:
                                 drv->status |= (drv->status & WD_DRQ) ? WD_LD : WD_DRQ;
@@ -673,7 +673,7 @@ static void wd1770_execute(wd1770_t *drv)
                             return;
                         }
                         drv->clk += BYTE_RATE;
-                        drv->data = (uint8_t)fdd_read(drv->fdd);
+                        drv->data = fdd_read(drv->fdd);
                         drv->status |= (drv->status & WD_DRQ) ? WD_LD : WD_DRQ;
                         continue;
                     case 7:
@@ -703,7 +703,7 @@ static void wd1770_execute(wd1770_t *drv)
                         }
                         drv->status |= (drv->status & WD_DRQ) ? WD_LD : WD_DRQ;
                         drv->clk += BYTE_RATE;
-                        drv->data = (uint8_t)fdd_read(drv->fdd);
+                        drv->data = fdd_read(drv->fdd);
                         if (drv->byte_count == 6) {
                             drv->sector = drv->data;
                         }
@@ -748,7 +748,7 @@ static void wd1770_execute(wd1770_t *drv)
                         drv->clk += BYTE_RATE;
                         res = drv->data;
                         if (drv->byte_count) {
-                            fdd_write(drv->fdd, (uint8_t)(drv->crc & 0xff));
+                            fdd_write(drv->fdd, drv->crc & 0xff);
                             drv->byte_count--;
                         } else {
                             drv->status |= (drv->status & WD_DRQ) ? WD_LD : WD_DRQ;
@@ -793,9 +793,9 @@ static void wd1770_execute(wd1770_t *drv)
                                 }
                             }
                             if (drv->tmp) {
-                                drv->crc = fdd_crc(drv->crc, (uint8_t)res);
+                                drv->crc = fdd_crc(drv->crc, res);
                             }
-                            fdd_write(drv->fdd, (uint8_t)res);
+                            fdd_write(drv->fdd, res);
                             drv->data = 0;
                         }
                         continue;
